Fix support to evaluate expressions with defaultable parameters
