+++ misc/debug-adapter/modules/debug-adapter-core/src/main/java/org/ballerinalang/debugadapter/evaluation/engine/invokable/GeneratedMethod.java
@@ -16,33 +16,14 @@
 
 package org.ballerinalang.debugadapter.evaluation.engine.invokable;
 
-import com.sun.jdi.AbsentInformationException;
-import com.sun.jdi.ClassNotLoadedException;
-import com.sun.jdi.DoubleType;
-import com.sun.jdi.FloatType;
-import com.sun.jdi.IntegerType;
-import com.sun.jdi.LocalVariable;
-import com.sun.jdi.LongType;
 import com.sun.jdi.Method;
-import com.sun.jdi.ShortType;
-import com.sun.jdi.Type;
 import com.sun.jdi.Value;
-import io.ballerina.runtime.api.types.BooleanType;
-import io.ballerina.runtime.api.types.ByteType;
 import org.ballerinalang.debugadapter.SuspendedContext;
 import org.ballerinalang.debugadapter.evaluation.EvaluationException;
 import org.ballerinalang.debugadapter.evaluation.EvaluationExceptionKind;
-import org.ballerinalang.debugadapter.evaluation.engine.Evaluator;
-import org.ballerinalang.debugadapter.evaluation.utils.EvaluationUtils;
-import org.ballerinalang.debugadapter.evaluation.utils.VMUtils;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-import static org.ballerinalang.debugadapter.evaluation.engine.InvocationArgProcessor.DEFAULTABLE_PARAM_SUFFIX;
-import static org.ballerinalang.debugadapter.evaluation.utils.EvaluationUtils.STRAND_VAR_NAME;
 
 /**
  * JVM generated method representation of a ballerina function.
@@ -61,25 +42,16 @@ public abstract class GeneratedMethod extends JvmMethod {
 
     @Override
     protected List<Value> getMethodArgs(JvmMethod method) throws EvaluationException {
-        if (argValues == null && argEvaluators == null) {
-            throw new EvaluationException(String.format(EvaluationExceptionKind.FUNCTION_EXECUTION_ERROR.getString()
-                    , methodRef.name()));
+        if (argValues == null) {
+            throw new EvaluationException(String.format(EvaluationExceptionKind.FUNCTION_EXECUTION_ERROR.getString(),
+                    methodRef.name()));
         }
 
         List<Value> argValueList = new ArrayList<>();
-        if (argValues != null) {
-            // Here we use the existing strand instance to execute the function invocation expression.
-            Value strand = context.getCurrentStrand();
-            argValueList.add(strand);
-            argValueList.addAll(argValues);
-            return argValueList;
-        }
-
-        // Evaluates all function argument expressions at first.
-        for (Map.Entry<String, Evaluator> argEvaluator : argEvaluators) {
-            argValueList.add(argEvaluator.getValue().evaluate().getJdiValue());
-        }
-
-        return EvaluationUtils.getAsObjects(context, argValueList);
+        // Here we use the existing strand instance to execute the function invocation expression.
+        Value strand = context.getCurrentStrand();
+        argValueList.add(strand);
+        argValueList.addAll(argValues);
+        return argValueList;
     }
 }
